add a separate file for test creation page
